No longer handle spells via E-ZRoller
With the way slot usage now works, we can't intercept it at this point.
Further, the damage spell level is set before the chat card is created,
and there's no good place in the workflow to re-inject/change this.

Really, macros are generally good enough, and aside from cantrip casting
this doesn't improve the DM QoL much anymore.